RBAC

+++ src/main/java/com/auth/backend_java/controller/TodoController.java
@@ -4,6 +4,7 @@ import com.auth.backend_java.model.Todo;
 import com.auth.backend_java.service.TodoService;
 import com.auth.backend_java.util.JwtUtil;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,6 +24,7 @@ public class TodoController {
 
 
     @PostMapping
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     public ResponseEntity<Todo> createTodo(
             @RequestBody Todo todo, 
             @RequestHeader("Authorization") String token) {
@@ -41,6 +43,7 @@ public class TodoController {
 
     // Get all Todos for a user
     @GetMapping
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     public ResponseEntity<List<Todo>> getTodos(@RequestHeader("Authorization") String token) {
         String username = jwtUtil.extractUsername(token.replace("Bearer", ""));
         List<Todo> todos = todoService.getTodos(username);
@@ -48,6 +51,7 @@ public class TodoController {
     }
 
     @PutMapping("/{id}")
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     public ResponseEntity<Todo> updateTodo(
             @PathVariable Long id, 
             @RequestBody Todo updatedTodo, 
@@ -59,12 +63,24 @@ public class TodoController {
     }
 
     @DeleteMapping("/{id}")
+    // @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     public ResponseEntity<Void> deleteTodo(
-            @PathVariable Long id, 
-            @RequestHeader("Authorization") String token) {
-        
-        String username = jwtUtil.extractUsername(token.replace("Bearer", ""));
-        todoService.deleteTodo(id, username);
-        return ResponseEntity.noContent().build();
+        @PathVariable Long id, 
+        @RequestHeader("Authorization") String token) {
+    
+
+    String role = jwtUtil.extractRole(token.replace("Bearer ", ""));
+    System.out.println("Role extracted for deletion: " + role);
+
+    if (!"ADMIN".equals(role)) {
+        throw new RuntimeException("Unauthorized: Only admins can delete todos.");
     }
+
+    
+    String username = jwtUtil.extractUsername(token.replace("Bearer ", ""));
+    
+   
+    todoService.deleteTodo(id, username);
+    return ResponseEntity.noContent().build();
+}
 }
